pyproject.toml
[project]
name = "marketing-agents"
version = "0.1.0"
description = "AI multi-agent for automated marketing content creation for Instagram and LinkedIn"
authors = [
     {name = "Triangle Media", email = "trianglemedia@tm.com"}
]
requires-python = ">=3.9"
dependencies = [
    "google-adk",
    "google-generativeai",
    "python-dotenv",
    "typer",
    "apscheduler",
    "pytz",
    "requests"
]

[tool.poetry.dependencies]
python = "^3.10"
google-genai = "*"
google-adk = ">=0.1.11"
python-dotenv = "*"
requests = "*"
typer = "^0.12.0"
apscheduler = "^3.10.4"
pytz = "*"

[tool.setuptools.packages.find]
where = ["src"]